PLAT-826 | Rename ApiSortableTable => PaginatedSortableTable
