+++ resources/assets/js/admin/components/lists/PaginatedList.vue
@@ -150,7 +150,7 @@ export default {
 		},
 		async dirty() {
 			if (this.dirty) {
-				this.fetch();
+				await this.fetch();
 			}
 
 			this.$emit('updated');
