+++ resources/assets/js/admin/components/lists/PaginatedSortableTable.vue
